12_Voice_RAG_Agent/12. Voice Rag/requirements.txt
openai
gradio
gTTS
faiss-cpu
python-dotenv
pyyaml
pydantic
numpy
